Make it utility on top of String
