--- tests/compiler/std/encoding.json
@@ -1,5 +1,0 @@
-{
-  "asc_flags": [
-    "--runtime none"
-  ]
-}
