+++ tests/compiler/std/string-encoding.json
@@ -1,0 +1,6 @@
+{
+  "asc_flags": [
+    "--runtime half",
+    "--use ASC_RTRACE=1"
+  ]
+}
