+++ tests/compiler/std/string-encoding.ts
@@ -5,12 +5,12 @@ var str = "𐐷hi𤭢";
 // UTF-16
 
 function testUTF16Length(): void {
-  assert(Encoding.UTF16.byteLength(str) == 12);
+  assert(String.UTF16.byteLength(str) == 12);
 }
 testUTF16Length();
 
 function testUTF16Encode(): void {
-  var buf = Encoding.UTF16.encode(str);
+  var buf = String.UTF16.encode(str);
   var ptr = changetype<usize>(buf);
   assert(buf.byteLength == 12);
   assert(load<u8>(ptr, 0) == 0x01);
@@ -29,36 +29,36 @@ function testUTF16Encode(): void {
 testUTF16Encode();
 
 function testUTF16Decode(): void {
-  var buf = Encoding.UTF16.encode(str);
-  assert(Encoding.UTF16.decode(buf) == str);
+  var buf = String.UTF16.encode(str);
+  assert(String.UTF16.decode(buf) == str);
 }
 testUTF16Decode();
 
 function testUTF16DecodeUnsafe(): void {
-  var buf = Encoding.UTF16.encode(str);
-  var len = Encoding.UTF16.byteLength(str);
+  var buf = String.UTF16.encode(str);
+  var len = String.UTF16.byteLength(str);
   var ptr = changetype<usize>(buf);
 
-  assert(Encoding.UTF16.decodeUnsafe(ptr, 0) == "");
-  assert(Encoding.UTF16.decodeUnsafe(ptr, len) == str);
-  assert(Encoding.UTF16.decodeUnsafe(ptr, 4) == "𐐷");
-  assert(Encoding.UTF16.decodeUnsafe(ptr + 4, 2) == "h");
-  assert(Encoding.UTF16.decodeUnsafe(ptr + 4, 4) == "hi");
-  assert(Encoding.UTF16.decodeUnsafe(ptr + 8, 4) == "𤭢");
-  assert(Encoding.UTF16.decodeUnsafe(ptr + 12, 0) == "");
+  assert(String.UTF16.decodeUnsafe(ptr, 0) == "");
+  assert(String.UTF16.decodeUnsafe(ptr, len) == str);
+  assert(String.UTF16.decodeUnsafe(ptr, 4) == "𐐷");
+  assert(String.UTF16.decodeUnsafe(ptr + 4, 2) == "h");
+  assert(String.UTF16.decodeUnsafe(ptr + 4, 4) == "hi");
+  assert(String.UTF16.decodeUnsafe(ptr + 8, 4) == "𤭢");
+  assert(String.UTF16.decodeUnsafe(ptr + 12, 0) == "");
 }
 testUTF16DecodeUnsafe();
 
 // UTF-8
 
 function testUTF8Length(): void {
-  assert(Encoding.UTF8.byteLength(str) == 10);
-  assert(Encoding.UTF8.byteLength(str, true) == 11);
+  assert(String.UTF8.byteLength(str) == 10);
+  assert(String.UTF8.byteLength(str, true) == 11);
 }
 testUTF8Length();
 
 function testUTF8Encode(): void {
-  var buf = Encoding.UTF8.encode(str);
+  var buf = String.UTF8.encode(str);
   var ptr = changetype<usize>(buf);
   assert(buf.byteLength == 10);
   assert(load<u8>(ptr, 0) == 0xF0);
@@ -75,7 +75,7 @@ function testUTF8Encode(): void {
 testUTF8Encode();
 
 function testUTF8EncodeNullTerminated(): void {
-  var buf = Encoding.UTF8.encode(str, true);
+  var buf = String.UTF8.encode(str, true);
   var ptr = changetype<usize>(buf);
   assert(buf.byteLength == 11);
   assert(load<u8>(ptr, 0) == 0xF0);
@@ -93,40 +93,40 @@ function testUTF8EncodeNullTerminated(): void {
 testUTF8EncodeNullTerminated();
 
 function testUTF8Decode(): void {
-  var buf = Encoding.UTF8.encode(str);
-  assert(Encoding.UTF8.decode(buf) == str);
+  var buf = String.UTF8.encode(str);
+  assert(String.UTF8.decode(buf) == str);
 }
 testUTF8Decode();
 
 function testUTF8DecodeNullTerminated(): void {
-  var buf = Encoding.UTF8.encode(str, true);
-  assert(Encoding.UTF8.decode(buf, true) == str);
+  var buf = String.UTF8.encode(str, true);
+  assert(String.UTF8.decode(buf, true) == str);
   var str2 = "123\0456";
-  assert(Encoding.UTF8.byteLength(str2, true) == 4);
-  var buf2 = Encoding.UTF8.encode(str2, true);
+  assert(String.UTF8.byteLength(str2, true) == 4);
+  var buf2 = String.UTF8.encode(str2, true);
   assert(buf2.byteLength == 4);
   var str3 = "123";
-  assert(Encoding.UTF8.byteLength(str3, true) == 4);
-  assert(Encoding.UTF8.decode(buf2, true) == str3);
-  var buf3 = Encoding.UTF8.encode(str2, false);
-  assert(Encoding.UTF8.decode(buf3, true) == str3);
+  assert(String.UTF8.byteLength(str3, true) == 4);
+  assert(String.UTF8.decode(buf2, true) == str3);
+  var buf3 = String.UTF8.encode(str2, false);
+  assert(String.UTF8.decode(buf3, true) == str3);
 }
 testUTF8DecodeNullTerminated();
 
 function testUTF8DecodeUnsafe(): void {
-  var buf = Encoding.UTF8.encode(str, true);
-  var len = Encoding.UTF8.byteLength(str, false); // !
+  var buf = String.UTF8.encode(str, true);
+  var len = String.UTF8.byteLength(str, false); // !
   var ptr = changetype<usize>(buf);
 
-  assert(Encoding.UTF8.decodeUnsafe(ptr, 0) == "");
-  assert(Encoding.UTF8.decodeUnsafe(ptr, len) == str);
-  assert(Encoding.UTF8.decodeUnsafe(ptr, 4) == "𐐷");
-  assert(Encoding.UTF8.decodeUnsafe(ptr + 4, 2) == "hi");
-  assert(Encoding.UTF8.decodeUnsafe(ptr + 6, 4) == "𤭢");
-  assert(Encoding.UTF8.decodeUnsafe(ptr + 10, 0) == "");
+  assert(String.UTF8.decodeUnsafe(ptr, 0) == "");
+  assert(String.UTF8.decodeUnsafe(ptr, len) == str);
+  assert(String.UTF8.decodeUnsafe(ptr, 4) == "𐐷");
+  assert(String.UTF8.decodeUnsafe(ptr + 4, 2) == "hi");
+  assert(String.UTF8.decodeUnsafe(ptr + 6, 4) == "𤭢");
+  assert(String.UTF8.decodeUnsafe(ptr + 10, 0) == "");
 
-  assert(Encoding.UTF8.decodeUnsafe(ptr + 4, 100, true) == "hi𤭢");
-  assert(Encoding.UTF8.decodeUnsafe(ptr + 6, 100, true) == "𤭢");
-  assert(Encoding.UTF8.decodeUnsafe(ptr + 10, 100, true) == "");
+  assert(String.UTF8.decodeUnsafe(ptr + 4, 100, true) == "hi𤭢");
+  assert(String.UTF8.decodeUnsafe(ptr + 6, 100, true) == "𤭢");
+  assert(String.UTF8.decodeUnsafe(ptr + 10, 100, true) == "");
 }
 testUTF8DecodeUnsafe();
