--- tests/compiler/std/string-utf8.json
@@ -1,5 +1,0 @@
-{
-  "asc_flags": [
-    "--runtime none"
-  ]
-}
